Fix issue 4
