+++ plugin.php
@@ -105,11 +105,11 @@ if (!class_exists('WPGraphQLGutenberg')) {
 			});
 
 			// Init Gutenberg ACF if WPGraphQLGutenbergACF plugin is not installed and GraphQLACF installed.
-			// add_action( 'init', function() {
-			// 	if ( ! class_exists( 'WPGraphQLGutenbergACF' ) && class_exists( 'WPGraphQL\ACF\Config' ) ) {
-			// 		\WPGraphQLGutenberg\Config::instance();
-			// 	}
-			// } );
+			add_action( 'acf/init', function() {
+				if ( ! class_exists( 'WPGraphQLGutenbergACF' ) && class_exists( 'WPGraphQL\ACF\Config' ) ) {
+					\WPGraphQLGutenberg\Config::instance();
+				}
+			} );
 
 			register_deactivation_hook(__FILE__, function () {
 				\WPGraphQLGutenberg\Server\Server::cleanup();
